feat: token update versioning
