+++ hedera-node/hedera-smart-contract-service-impl/src/main/java/com/hedera/node/app/service/contract/impl/exec/systemcontracts/hss/schedulenative/ScheduleNativeTranslator.java
@@ -18,7 +18,7 @@ package com.hedera.node.app.service.contract.impl.exec.systemcontracts.hss.sched
 
 import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.HtsSystemContract.HTS_167_EVM_ADDRESS;
 import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.create.CreateTranslator.createMethodsMap;
-import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.UpdateTranslator.updateMethodsMap;
+import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.UpdateTokenCommons.updateMethodsSet;
 import static com.hedera.node.app.spi.workflows.HandleException.validateTrue;
 import static java.util.Objects.requireNonNull;
 import static org.hyperledger.besu.datatypes.Address.fromHexString;
@@ -135,7 +135,7 @@ public class ScheduleNativeTranslator extends AbstractCallTranslator<HssCallAtte
         final var canBeCreateToken =
                 createMethodsMap.keySet().stream().anyMatch(s -> Arrays.equals(s.selector(), innerCallSelector));
         final var canBeUpdateToken =
-                updateMethodsMap.keySet().stream().anyMatch(s -> Arrays.equals(s.selector(), innerCallSelector));
+                updateMethodsSet.stream().anyMatch(s -> Arrays.equals(s.selector(), innerCallSelector));
 
         return canBeCreateToken || canBeUpdateToken;
     }
