+++ hedera-node/hedera-smart-contract-service-impl/src/main/java/com/hedera/node/app/service/contract/impl/exec/systemcontracts/hts/update/UpdateTokenCommonTranslator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023-2025 Hedera Hashgraph, LLC
+ * Copyright (C) 2025 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,10 +16,6 @@
 
 package com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update;
 
-import static com.hedera.node.app.hapi.utils.contracts.ParsingConstants.EXPIRY;
-import static com.hedera.node.app.hapi.utils.contracts.ParsingConstants.EXPIRY_V2;
-import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.UpdateDecoder.FAILURE_CUSTOMIZER;
-
 import com.hedera.hapi.node.base.AccountID;
 import com.hedera.hapi.node.transaction.TransactionBody;
 import com.hedera.node.app.service.contract.impl.exec.gas.DispatchType;
@@ -29,63 +25,36 @@ import com.hedera.node.app.service.contract.impl.exec.systemcontracts.common.Abs
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.common.Call;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.DispatchForResponseCodeHtsCall;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.HtsCallAttempt;
-import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.ReturnTypes;
+import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.address_0x167.UpdateTranslator;
 import com.hedera.node.app.service.contract.impl.exec.utils.SystemContractMethod;
-import com.hedera.node.app.service.contract.impl.exec.utils.SystemContractMethod.Category;
-import com.hedera.node.app.service.contract.impl.exec.utils.SystemContractMethod.Variant;
 import com.hedera.node.app.service.contract.impl.exec.utils.SystemContractMethodRegistry;
 import com.hedera.node.app.service.contract.impl.hevm.HederaWorldUpdater;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
-import javax.inject.Inject;
-import javax.inject.Singleton;
 
-/**
- * Translates ERC-721 {@code updateTokenExpiryInfo()} calls to the HTS system contract.
- */
-@Singleton
-public class UpdateExpiryTranslator extends AbstractCallTranslator<HtsCallAttempt> {
-    /**
-     * Selector for updateTokenExpiryInfo(address, EXPIRY) method.
-     */
-    public static final SystemContractMethod UPDATE_TOKEN_EXPIRY_INFO_V1 = SystemContractMethod.declare(
-                    "updateTokenExpiryInfo(address," + EXPIRY + ")", ReturnTypes.INT)
-            .withVariant(Variant.V1)
-            .withCategories(Category.UPDATE);
-    /**
-     * Selector for updateTokenExpiryInfo(address, EXPIRY_V2) method.
-     */
-    public static final SystemContractMethod UPDATE_TOKEN_EXPIRY_INFO_V2 = SystemContractMethod.declare(
-                    "updateTokenExpiryInfo(address," + EXPIRY_V2 + ")", ReturnTypes.INT)
-            .withVariant(Variant.V2)
-            .withCategories(Category.UPDATE);
+public class UpdateTokenCommonTranslator extends AbstractCallTranslator<HtsCallAttempt> {
 
-    private final UpdateDecoder decoder;
+    protected static final String UPDATE_TOKEN_INFO_STRING = "updateTokenInfo(address,";
 
-    /**
-     * @param decoder the decoder used for token update calls
-     */
-    @Inject
-    public UpdateExpiryTranslator(
-            UpdateDecoder decoder,
+    protected final Map<SystemContractMethod, UpdateDecoderFunction> updateMethodsMap = new HashMap<>();
+
+    public UpdateTokenCommonTranslator(
             @NonNull final SystemContractMethodRegistry systemContractMethodRegistry,
             @NonNull final ContractMetrics contractMetrics) {
         super(SystemContractMethod.SystemContract.HTS, systemContractMethodRegistry, contractMetrics);
-        this.decoder = decoder;
-
-        registerMethods(UPDATE_TOKEN_EXPIRY_INFO_V1, UPDATE_TOKEN_EXPIRY_INFO_V2);
     }
 
     @Override
     public @NonNull Optional<SystemContractMethod> identifyMethod(@NonNull final HtsCallAttempt attempt) {
-        return attempt.isMethod(UPDATE_TOKEN_EXPIRY_INFO_V1, UPDATE_TOKEN_EXPIRY_INFO_V2);
+        return updateMethodsMap.keySet().stream().filter(attempt::isSelector).findFirst();
     }
 
     @Override
     public Call callFrom(@NonNull HtsCallAttempt attempt) {
         return new DispatchForResponseCodeHtsCall(
-                attempt, nominalBodyFor(attempt), UpdateExpiryTranslator::gasRequirement, FAILURE_CUSTOMIZER);
+                attempt, nominalBodyFor(attempt), UpdateTranslator::gasRequirement, UpdateDecoder.FAILURE_CUSTOMIZER);
     }
 
     /**
@@ -104,10 +73,10 @@ public class UpdateExpiryTranslator extends AbstractCallTranslator<HtsCallAttemp
     }
 
     private TransactionBody nominalBodyFor(@NonNull final HtsCallAttempt attempt) {
-        if (Arrays.equals(attempt.selector(), UPDATE_TOKEN_EXPIRY_INFO_V1.selector())) {
-            return decoder.decodeTokenUpdateExpiryV1(attempt);
-        } else {
-            return decoder.decodeTokenUpdateExpiryV2(attempt);
-        }
+        return updateMethodsMap.entrySet().stream()
+                .filter(entry -> attempt.isSelector(entry.getKey()))
+                .map(entry -> entry.getValue().decode(attempt))
+                .findFirst()
+                .orElse(null);
     }
 }
