+++ hedera-node/hedera-smart-contract-service-impl/src/main/java/com/hedera/node/app/service/contract/impl/exec/systemcontracts/hts/update/UpdateTokenCommons.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (C) 2025 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update;
+
+import com.hedera.node.app.service.contract.impl.exec.utils.SystemContractMethod;
+import java.util.HashSet;
+import java.util.Set;
+import javax.inject.Singleton;
+
+@Singleton
+public final class UpdateTokenCommons {
+
+    public static final Set<SystemContractMethod> updateMethodsSet = new HashSet<>();
+
+    private UpdateTokenCommons() {}
+}
