+++ hedera-node/hedera-smart-contract-service-impl/src/test/java/com/hedera/node/app/service/contract/impl/test/exec/systemcontracts/hts/update/address_0x16c/UpdateTranslatorTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023-2025 Hedera Hashgraph, LLC
+ * Copyright (C) 2025 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,23 +14,27 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.hts.burn;
+package com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.hts.update.address_0x16c;
 
-import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.burn.BurnTranslator.BURN_TOKEN_V1;
-import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.burn.BurnTranslator.BURN_TOKEN_V2;
-import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.address_0x167.UpdateTranslator.TOKEN_UPDATE_INFO_FUNCTION_V3;
-import static com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.CallAttemptHelpers.prepareHtsAttemptWithSelector;
+import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.HtsSystemContract.HTS_167_CONTRACT_ID;
+import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.HtsSystemContract.HTS_16C_CONTRACT_ID;
+import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.address_0x16c.UpdateTranslator.TOKEN_UPDATE_INFO_FUNCTION_WITH_METADATA;
+import static com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.CallAttemptHelpers.prepareHtsAttemptWithSelectorWithContractID;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.hedera.node.app.service.contract.impl.exec.gas.SystemContractGasCalculator;
 import com.hedera.node.app.service.contract.impl.exec.metrics.ContractMetrics;
 import com.hedera.node.app.service.contract.impl.exec.scope.VerificationStrategies;
+import com.hedera.node.app.service.contract.impl.exec.scope.VerificationStrategy;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.AddressIdConverter;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.HtsCallAttempt;
-import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.burn.BurnDecoder;
-import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.burn.BurnTranslator;
+import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.freeze.FreezeUnfreezeTranslator;
+import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.UpdateDecoder;
+import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.address_0x16c.UpdateTranslator;
 import com.hedera.node.app.service.contract.impl.exec.utils.SystemContractMethodRegistry;
 import com.hedera.node.app.service.contract.impl.hevm.HederaWorldUpdater;
+import com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.common.CallTestBase;
+import com.hedera.node.app.service.token.ReadableAccountStore;
+import com.hedera.node.app.service.token.ReadableTokenStore;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -38,41 +42,48 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class BurnTranslatorTest {
+class UpdateTranslatorTest extends CallTestBase {
 
     @Mock
     private HtsCallAttempt attempt;
 
     @Mock
-    private SystemContractGasCalculator gasCalculator;
+    private AddressIdConverter addressIdConverter;
 
     @Mock
-    private AddressIdConverter addressIdConverter;
+    private VerificationStrategy verificationStrategy;
+
+    @Mock
+    private VerificationStrategies verificationStrategies;
 
     @Mock
     private HederaWorldUpdater.Enhancement enhancement;
 
     @Mock
-    private VerificationStrategies verificationStrategies;
+    private ReadableTokenStore readableTokenStore;
+
+    @Mock
+    private ReadableAccountStore readableAccountStore;
 
     @Mock
     private ContractMetrics contractMetrics;
 
     private final SystemContractMethodRegistry systemContractMethodRegistry = new SystemContractMethodRegistry();
 
-    private BurnTranslator subject;
+    private UpdateTranslator subject;
 
-    private final BurnDecoder decoder = new BurnDecoder();
+    private final UpdateDecoder decoder = new UpdateDecoder();
 
     @BeforeEach
     void setUp() {
-        subject = new BurnTranslator(decoder, systemContractMethodRegistry, contractMetrics);
+        subject = new UpdateTranslator(decoder, systemContractMethodRegistry, contractMetrics);
     }
 
     @Test
-    void matchesBurnTokenV1() {
-        attempt = prepareHtsAttemptWithSelector(
-                BURN_TOKEN_V1,
+    void matchesUpdateMetadataTest() {
+        attempt = prepareHtsAttemptWithSelectorWithContractID(
+                HTS_16C_CONTRACT_ID,
+                TOKEN_UPDATE_INFO_FUNCTION_WITH_METADATA,
                 subject,
                 enhancement,
                 addressIdConverter,
@@ -83,22 +94,10 @@ class BurnTranslatorTest {
     }
 
     @Test
-    void matchesBurnTokenV2() {
-        attempt = prepareHtsAttemptWithSelector(
-                BURN_TOKEN_V2,
-                subject,
-                enhancement,
-                addressIdConverter,
-                verificationStrategies,
-                gasCalculator,
-                systemContractMethodRegistry);
-        assertThat(subject.identifyMethod(attempt)).isPresent();
-    }
-
-    @Test
-    void matchFailsOnInvalidSelector() {
-        attempt = prepareHtsAttemptWithSelector(
-                TOKEN_UPDATE_INFO_FUNCTION_V3,
+    void matchesFailsOnIncorrectSelector() {
+        attempt = prepareHtsAttemptWithSelectorWithContractID(
+                HTS_167_CONTRACT_ID,
+                FreezeUnfreezeTranslator.FREEZE,
                 subject,
                 enhancement,
                 addressIdConverter,
